requirements.txt
python-telegram-bot==22.7
httpx==0.27.0
requests==2.32.3
